// frontend/src/components/sign-up/types.d.ts
export type PersonalDataTypes = {
  name: string;
  email: string;
  password: string;
  password2: string;
  bio: string;
};

export type RoleDataTypes = {
  role: number;
};

export type SocialDataTypes = {
  social_twitter: string;
  social_telegram: string;
  social_instagram: string;
  social_discord: string;
};

export type ProjectDataTypes = {
  project_name: string;
  project_website: string;
  project_telegram: string;
  project_twitter: string;
  project_instagram: string;
  project_discord: string;
  project_logo: string;
};